migrate packages (#17)
add a keyword argument migrate_packages that is true by default whenever there is 
- not already an environment for the target version and
- the update is an upgrade according to prefer

Justification for the second point: if someone is updating to a lesser preferred version, they are either trying something on an old version of Julia or an experimental version. In either case, they likely do not want a pre-populated global environment. And if they do, it is easier to run again with migrate_packages = :true than to manually delete the environment in .julia.